Added: - maskedmax utility function -  unity element not well defined**!!
Fixed:
- bug where individuals=/=individualsss
-   Soln: truncate psis_i and use maskedmax
-   Details:

Individuals was giving max 0.35,.. at [11][0], individualsss was giving 0:
In individuals we wanted to broadcast (arb_i) tensor across a (arb_i,arb_j) to get an (arb_i,arb_j) tensor, however tensor a was actually a (7) tensor due to padding, it just so happened that arb_i was 1 and so the operation (1,7).T+(7) could succeed, resulting in a (7,7) tensor due to broadcasting in the wrong dimension.
This was due to psis_i in individuals being a (7) not an (arb_i) tensor
Fixing this (by cutting psis_i accordingly) we get max -0.02... at [11][0] according to individuals, this appears correct. To fix this in individualsss we take the max w.r.t the mask.

Test now passes (both give -0.02...)