!!!DEBUGGING!!!
Problem identified:
smartsum may require different behaviour:
When summing phi_k over k we're summing [nan,nan,nan] if there is no candidate for i or j, in this case phi should be nan as there is no candidate - the nans represent a lack of candidate for the i,j indices, not the k(summing) index.
However there may be times when summing [nan,nan,nan] which represent the summing index, in which case the result should be a default of 0? (t.b.c.)